# dist/tsoa/swagger.json
{
    "components": {
...
        "schemas": {
...
            "CustomerRequestParameter": {
                "properties": {
                    "piutang": {
                        "type": "number",
                        "format": "double"
                    },
                    "hutang": {
                        "type": "number",
                        "format": "double"
                    },
                    "contact_number": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "contact_number",
                    "name"
                ],
                "type": "object"
            },
            "CustomerUpdateRequestParameter": {
                "properties": {
                    "piutang": {
                        "type": "number",
                        "format": "double"
                    },
                    "hutang": {
                        "type": "number",
                        "format": "double"
                    },
                    "contact_number": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateEmployeeRequest": {
                "properties": {
                    "role_id": {
                        "type": "number",
                        "format": "double"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "noInduk": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "role_id",
                    "phone_number",
                    "birth_date",
                    "noInduk",
                    "email",
                    "name"
                ],
                "type": "object"
            },
            "ChangeEmployeeRole": {
                "properties": {
                    "role_id": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "role_id"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "access_token",
                "in": "header",
                "flow": "implicit",
                "scopes": {
                    "write:pets": "modify things",
                    "read:pets": "read things"
                }
            }
        }
    },
    "info": {
        "title": ""
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/customer": {
            "get": {
                "operationId": "GetAllCustomer",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Customer"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:customer"
                        ]
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "CreateCustomer",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Customer"
                ],
                "security": [
                    {
                        "api_key": [
                            "create:customer"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerRequestParameter"
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/detail/{id}": {
            "get": {
                "operationId": "FindCustomerById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Customer"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:customer"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/customer/search": {
            "get": {
                "operationId": "SearchCustomer",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Customer"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:customer"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/customer/{id}": {
            "put": {
                "operationId": "UpdateCustomer",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Customer"
                ],
                "security": [
                    {
                        "api_key": [
                            "update:customer"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerUpdateRequestParameter"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteCustomer",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Customer"
                ],
                "security": [
                    {
                        "api_key": [
                            "delete:customer"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/pegawai": {
            "get": {
                "operationId": "GetAllEmployee",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Pegawai"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:pegawai"
                        ]
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "CreateEmployee",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Pegawai"
                ],
                "security": [
                    {
                        "api_key": [
                            "create:pegawai"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateEmployeeRequest"
                            }
                        }
                    }
                }
            }
        },
        "/api/pegawai/{id}": {
            "put": {
                "operationId": "UpdateEmployee",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Pegawai"
                ],
                "security": [
                    {
                        "api_key": [
                            "update:pegawai"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateEmployeeRequest"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteEmployee",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Pegawai"
                ],
                "security": [
                    {
                        "api_key": [
                            "delete:pegawai"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "patch": {
                "operationId": "ChangeEmployeeRole",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Pegawai"
                ],
                "security": [
                    {
                        "api_key": [
                            "update:pegawai"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeEmployeeRole"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
}